Migrate interface declarations to agent setup and connect proxy. Migrate telemetry config to its own file. Add docs and field various review items
